* some cleanup in gx_amp.lv2

+++ trunk/src/LV2/gx_amp.lv2/DSP/cab_data.cc
@@ -820,7 +820,7 @@ struct CabEntry {
     { &static_cast<CabDesc&>(cab_data_vitalize) },
     { &static_cast<CabDesc&>(cab_data_charisma) },
 };
-static const uint32_t cab_table_size = sizeof(cab_table) / sizeof(cab_table[0]);
+static const size_t cab_table_size = sizeof(cab_table) / sizeof(cab_table[0]);
 
 static CabEntry& getCabEntry(uint32_t n) {
     if (n >= cab_table_size) {
