+++ trunk/src/LV2/gx_amp.lv2/DSP/gx_amp.h
@@ -69,7 +69,7 @@ static plug amp_model [] = {
     gxamp6::plugin, //17
 };
 
-static const uint32_t AMP_COUNT = sizeof(amp_model) / sizeof(amp_model[0]);
+static const size_t AMP_COUNT = sizeof(amp_model) / sizeof(amp_model[0]);
 
 
 #endif  //SRC_HEADERS_GX_AMP_H_
